Extracted specific currency related method to separate module
